Revert "Add .shadow-cljs cache to :clean-targets"
This reverts commit 2130268c97ec2302c847906025046f5377f48444.

Re @thheller's feedback to please not ever wipe this dir.

Caching is supposed to be reliable and it is. If you actually have to
delete this folder to fix something he would much rather hear about it
and fix the issue than having lein clean as the default go to.

Co-Authored-By: Thomas Heller <th.heller@gmail.com>